+++ PowerEditor/src/ScitillaComponent/FunctionCallTip.h
@@ -59,8 +59,8 @@ private:
 	stringVec _retVals;				//vector of overload return values/types
 	std::vector<stringVec> _overloads;	//vector of overload params (=vector)
 	stringVec _descriptions;		//vecotr of function descriptions
-	int _currentNrOverloads = 0;		//current amount of overloads
-	int _currentOverload = 0;			//current chosen overload
+	size_t _currentNrOverloads = 0;		//current amount of overloads
+	size_t _currentOverload = 0;			//current chosen overload
 	int _currentParam = 0;				//current highlighted param
 
 	TCHAR _start = '(';
